backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "color-restoration"
version = "0.1.0"
description = "color-restoration — görüntü renklendirme API ve servisleri"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "pillow>=10.4.0",
  "numpy>=1.26.0",
  "tensorflow>=2.16.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["train*", "models*", "scripts*", "tests*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
